Move frontend code to client subdirectory
