+++ src/index.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 
-import App from "./App";
-import "./index.css";
-import store from "./store";
+import App from "./client/App";
+import "./client/index.css";
+import store from "./client/store";
 
 ReactDOM.render(
   <Provider store={store}>
